# Fix index dimension mismatch in `TensorNetwork` construction