Closes #155