@Todorbsc 2024-10-11
@jofrevalles @mofeing the dimension mismatch seems to be fixed with the verification I implemented in `TensorNetwork` construction. I "uncommented" the TN dimension mismatch construction test and now it raises the proper exception as well as I tested the `Chain` example that Jofre commented in the [issue](https://github.com/bsc-quantic/Tenet.jl/issues/155#issue-2362798480), what raises an error now:
```julia
julia> mps = Chain(State(), Open(), [rand(2, 3), rand(2, 6, 4), rand(2, 5)])
ERROR: DimensionMismatch: Index D has inconsistent dimension: [3, 6]
Stacktrace:
 [1] TensorNetwork(tensors::Vector{Tensor{Float64, N, A} where {N, A<:AbstractArray{Float64, N}}})
   @ Tenet ~/QuanticGit/Tenet.jl/src/TensorNetwork.jl:58
 [2] Chain(::State, boundary::Open, arrays::Vector{Array{Float64}}; order::Tuple{Symbol, Symbol, Symbol})
   @ Tenet ~/QuanticGit/Tenet.jl/src/Ansatz/Chain.jl:99
 [3] Chain(::State, boundary::Open, arrays::Vector{Array{Float64}})
   @ Tenet ~/QuanticGit/Tenet.jl/src/Ansatz/Chain.jl:63
 [4] top-level scope
   @ REPL[3]:1
```
Also, let me know if there's something else needed, specially with what you commented in the issue about the `unsafe_region`.