@mofeing 2024-10-11
Do you mind benchmarking with and without safe guards?