@Todorbsc 2024-10-11
> Do you mind benchmarking with and without the safe guards / checks for some large TNs? Like 100-1000 tensors

I benchmarked for `rand(TensorNetwork, 1000, 3)`.

Without the indices safe region:
```julia
julia> @benchmark tn = rand(TensorNetwork, 1000, 3)
BenchmarkTools.Trial: 88 samples with 1 evaluation.
 Range (min … max):  17.460 ms … 560.275 ms  ┊ GC (min … max):  0.00% … 12.05%
 Time  (median):     25.241 ms               ┊ GC (median):     2.20%
 Time  (mean ± σ):   56.874 ms ±  82.945 ms  ┊ GC (mean ± σ):  22.67% ± 24.72%

  █▅▁                                                           
  █████▅▁▅▁▁▅▆▆▅▇▅▁▁▅▁▆▁▁▁▁▁▅▁▁▁▁▁▁▁▁▁▅▁▁▁▅▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▅ ▁
  17.5 ms       Histogram: log(frequency) by time       407 ms <

 Memory estimate: 25.09 MiB, allocs estimate: 64188.
 ```
 
 With the indices safe region:
 ```julia
julia> @benchmark tn = rand(TensorNetwork, 1000, 3)
BenchmarkTools.Trial: 89 samples with 1 evaluation.
 Range (min … max):  23.384 ms … 505.261 ms  ┊ GC (min … max):  0.00% … 15.44%
 Time  (median):     30.132 ms               ┊ GC (median):     0.00%
 Time  (mean ± σ):   61.098 ms ±  85.135 ms  ┊ GC (mean ± σ):  12.13% ± 14.75%

  █▃ ▁                                                          
  ████▇█▁▅▁▇▁▁▁▆▁▁▆▅▁▁▁▁▁▁▁▁▁▁▁▅▅▁▁▁▁▁▁▁▁▁▁▅▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▅ ▁
  23.4 ms       Histogram: log(frequency) by time       485 ms <

 Memory estimate: 28.02 MiB, allocs estimate: 103174.
 ```