@Todorbsc 2024-10-11
@mofeing Another benchmark for `rand(TensorNetwork, 200, 3)`.

Using checks:
```julia
julia> @benchmark tn = rand(TensorNetwork, 200, 3)
BenchmarkTools.Trial: 400 samples with 1 evaluation.
 Range (min … max):   2.218 ms … 253.500 ms  ┊ GC (min … max):  0.00% … 23.87%
 Time  (median):      3.163 ms               ┊ GC (median):     0.00%
 Time  (mean ± σ):   12.499 ms ±  28.164 ms  ┊ GC (mean ± σ):  31.62% ± 21.23%

  █▄▁                                                           
  ███▇▅▆▅▅▅▄▅▅▅▅▄▅▁▅▁▁▄▄▁▁▁▄▁▁▁▄▄▆▁▁▅▆▁▅▁▄▁▄▄▁▄▁▅▁▁▁▁▁▁▁▄▁▄▁▄▄ ▆
  2.22 ms       Histogram: log(frequency) by time       129 ms <

 Memory estimate: 2.80 MiB, allocs estimate: 19763.
 ```
 
 Wiithout checks:
 ```julia
julia> @benchmark tn = rand(TensorNetwork, 200, 3)
BenchmarkTools.Trial: 353 samples with 1 evaluation.
 Range (min … max):   1.240 ms …   1.031 s  ┊ GC (min … max):  0.00% …  0.32%
 Time  (median):      2.028 ms              ┊ GC (median):     0.00%
 Time  (mean ± σ):   14.150 ms ± 65.195 ms  ┊ GC (mean ± σ):  30.34% ± 22.77%

  █                                                            
  ██▅▆▄▅▅▁▁▄▁▁▁▁▄▅▅▅▄▆▄▆▁▁▁▄▁▄▁▁▁▁▄▁▁▁▁▁▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄ ▅
  1.24 ms      Histogram: log(frequency) by time       222 ms <

 Memory estimate: 2.84 MiB, allocs estimate: 11968.
 ```
 
 Min. time: Checks `2.218 ms` vs. No checks `1.240 ms`.