@mofeing 2024-10-11
Thanks!

> Correct me if I am wrong, but as regularity may influence the times, I would just look at the minimum time: Without the checks 17.460 ms vs. With the checks 23.384 ms.

Well, the min time can also be affected by jitter so I would take the mean which means there is a 20% increase in runtime... but it's still 25ms which is ok.